- removed media type control from collections - fixed 64-Bit issue with playlist index

+++ HoccerXO/Models/CollectionItem.h
@@ -15,6 +15,6 @@
 
 @property (nonatomic, retain) Attachment *attachment;
 @property (nonatomic, retain) Collection *collection;
-@property (nonatomic, assign) NSInteger   index;
+@property (nonatomic, assign) int        index;
 
 @end
